requirements.txt
transformers
faiss-cpu
streamlit
pandas